+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## [Unreleased](https://github.com/dalance/procs/compare/v0.9.19...Unreleased) - ReleaseDate
+## [Unreleased](https://github.com/dalance/procs/compare/v0.9.20...Unreleased) - ReleaseDate
+
+## [v0.9.20](https://github.com/dalance/procs/compare/v0.9.19...v0.9.20) - 2020-03-13
 
 * [Added] Tree color config [#50](https://github.com/dalance/procs/issues/50)
 * [Added] Black color and style [#49](https://github.com/dalance/procs/issues/49)